+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+# v1.0.1
+
+### Bugfix
+
+ - Reuse ununsed navigation container (#139)
+ - Make the CMS content dependant on the language and gender selection (#130)
+
+### Changes
+
+ - Add a style description for each style and generate the style list for each
+   Style group.
+ - Make CMS page and section navigation sticky.
+
+### New Features
+
+ - Allow to enable/disable the navigation menu on navigation pages (#75)
+
+------------
+
 # v1.0.0
 
 ### Bugfix
@@ -11,13 +30,13 @@
    no female version was available. Now it defaults back to the male label.
  - User inputs are now properly checked in the style `fromUserInput`.
 
-### Changes
-
 ### New Features
 
  - Improve the style `progressBar` (enable/disable label and stripes)
  - Improve the style `carousel` (enable/disable crossfade, allow captions)
  - Allow custom input fields in the style `validate` (#124).
+
+------------
 
 # v0.9.13
 
@@ -42,6 +61,8 @@
  - Allow to overwrite existing asset and css files
  - Allow to remove asset and css files from the server
  - A new style `carousel` to display images in a slide show (#125)
+
+------------
 
 # v0.9.12
 
@@ -70,6 +91,8 @@
      users from a room
    - A changed chat interface where rooms can be selected if available.
 
+------------
+
 # v0.9.11
 
 ### Bugfix
@@ -87,6 +110,8 @@
  - Allow to modify email content in the admin section (#141)
  - Implement automatic email reminder (#79)
 
+------------
+
 # v0.9.10
 
 ### Bugfix
@@ -96,6 +121,8 @@
 ### Changes
 
  - Update style descriptions (#138)
+
+------------
 
 # v0.9.9
 
@@ -113,12 +140,16 @@
    true (#128)
  - New style audio (#116)
 
+------------
+
 # v0.9.8
 
 ### New Features
 
  - Navigation pages can be added to the header menu (links to the first entry)
    (#134)
+
+------------
 
 # v0.9.7
 
@@ -140,6 +171,8 @@
  - A list of validation codes can be generated which can be used to register
    (#121)
 
+------------
+
 # v0.9.6
 
 ### Bugfix
@@ -160,6 +193,8 @@
    `nav\_css` (#105)
  - Allow to generate links to nav sections (#104)
 
+------------
+
 # v0.9.5
 
 ### Bugfix
@@ -173,6 +208,8 @@
 ### New Features
 
  - A new style `showUserInput` to display user input (#87)
+
+------------
 
 # v0.9.4
 
@@ -189,6 +226,8 @@
 
  - Style descriptions and style groups (#99)
  - Load default values of style fields from the db (#98)
+
+------------
 
 # v0.9.3